requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
gitdb==4.0.11
gitpython==3.1.40
idna==3.6
importlib-metadata==6.11.0
jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
mdurl==0.1.2
protobuf==4.25.1
pyarrow==14.0.2
pydeck==0.8.1b0
pymupdf==1.23.8
PyMuPDFb==1.23.7
PyPDF2==3.0.1
python_decouple==3.8
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.16.2
smmap==5.0.1
streamlit==1.28.0
streamlit_card==1.0.0
st-annotated-text==4.0.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0